requirements.txt
python-dotenv==0.19.0
python-telegram-bot==13.5
requests==2.26.0
